Moved Biometry functions to SharedSecurityManager
